# openapi/container-contract-v1.yaml
openapi: 3.1.0
info:
  title: Synth Containers Contract
  version: 0.2.0
  description: |
    Canonical public wire contract for the standardized Synth container runtime.
    It covers metadata discovery, task discovery, blocking and async rollouts,
    rollout lifecycle polling, checkpoints, resume semantics, and control
    surfaces such as pause and terminate.

    v0.2.0 (2026-05-28): RolloutPolicySpec is now a TYPED OpenAI-compatible
    inference target (the "openai contract"), superseding the prior free-form
    {config} bag and the per-container POLICY_* env side-channel. The GEPA
    optimizer sub-contract bumps to `synth_optimizers.gepa.v2`. Credentials are
    referenced via `credential_mode`; raw keys are never carried on the wire.
servers:
  - url: http://localhost
    description: Local development
paths:
...
  /rollouts/{rollout_id}/annotations/events:
    get:
      operationId: getRolloutAnnotationEvents
      summary: Read the provisional live annotation stream a bound protocol published beside the rollout
      description: >-
        Declared by the rollout stream descriptor as `annotation.events` when a live
        annotation protocol revision was pinned at prepare/start; never guess this route.
        Same envelope and cursor semantics as `/rollouts/{rollout_id}/events`, own
        sequence space. Rows carry `rollout_id` and `stream_id`. Findings are
        `provisional`; they never alter reward, achievements, terminal status or the
        sealed trace. See docs/specs/live-annotation-protocol-v1.md.
      parameters:
        - $ref: '#/components/parameters/RolloutId'
        - in: query
          name: after
          schema:
            type: integer
            minimum: 0
        - in: query
          name: limit
          schema:
            type: integer
            minimum: 1
            maximum: 10000
      responses:
        '200':
          description: Annotation event page (synth.live-annotation-stream.v1)
          content:
            application/json:
              schema:
                type: object
                properties:
                  schema:
                    type: string
                    enum: [synth.live-annotation-stream.v1]
                  rollout_id:
                    type: string
                  stream_id:
                    type: string
                  cursor:
                    type: object
                    additionalProperties: true
                  events:
                    type: array
                    items:
                      type: object
                      additionalProperties: true
                  summary:
                    type: object
                    nullable: true
                    additionalProperties: true
                additionalProperties: true
        '404':
          description: No protocol was bound to this rollout
  /rollouts/{rollout_id}/annotations/control:
    post:
      operationId: postRolloutAnnotationControl
      summary: Consumer -> annotator control for a running live annotation protocol
      description: >-
        Declared by the rollout stream descriptor as `annotation.control`. Ops:
        `message` (delivered to the protocol's on_message), `protocol.update`
        (hot-swap to another installed revision, carrying snapshot state when the
        protocol supports it), `stop`. The immediate answer is the sender's
        acknowledgement; the durable acknowledgement is published on the annotation
        stream as annotation.control.received / annotation.control.refused. The
        same JSON may be sent on the duplex `annotation.websocket`.
      parameters:
        - $ref: '#/components/parameters/RolloutId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [op]
              properties:
                schema:
                  type: string
                  enum: [synth.live-annotation-control.v1]
                op:
                  type: string
                  enum: [message, protocol.update, stop]
                control_id:
                  type: string
                message:
                  type: object
                  additionalProperties: true
                protocol_revision_id:
                  type: string
                carry_state:
                  type: boolean
                reason:
                  type: string
      responses:
        '202':
          description: Accepted and queued for the annotator
        '422':
          description: Refused (contract breach, unknown revision, or sealed stream)
        '404':
          description: No annotation stream for this rollout
        '409':
          description: The annotator is not active for this rollout
  /annotation-protocol:
    get:
      operationId: getAnnotationProtocol
      summary: Installed live annotation protocol identity (never source or credentials)
      responses:
        '200':
          description: synth.container-annotation-protocol.v1
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
    put:
      operationId: putAnnotationProtocol
      summary: Install a digest-pinned live annotation protocol revision
      description: >-
        Mirrors `PUT /policy`: content-addressed `anprev_<sha256[:16]>` revision over
        code, protocol_id, configuration and source_revision; idempotent on identical
        inputs; refuses credential-shaped configuration keys and code that cannot boot
        in an isolated stdlib-only process.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [code, protocol_id]
              properties:
                code:
                  type: string
                protocol_id:
                  type: string
                configuration:
                  type: object
                  additionalProperties: true
                source_revision:
                  type: string
                  nullable: true
      responses:
        '200':
          description: Installed identity plus idempotency flag and isolation receipt
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
        '422':
          description: protocol_identity_required, protocol_source_required, protocol_credential_forbidden, protocol_boot_failed, or protocol_identity_mismatch
...
components:
  parameters:
    RolloutId:
      name: rollout_id
      in: path
      required: true
      schema:
        type: string
...